pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chroma>=0.2.0",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.2",
]
